Moved MySQL\QueryObject into Language\SQL namespace
