+++ src/Storage/Database/Language/SQL/Query/Update.php
@@ -10,7 +10,7 @@ use Projom\Storage\Database\Language\SQL\Table;
 use Projom\Storage\Database\Language\SQL\QueryInterface;
 use Projom\Storage\Database\Language\SQL\Join;
 use Projom\Storage\Database\Language\SQL\Util;
-use Projom\Storage\Database\MySQL\QueryObject;
+use Projom\Storage\Database\Language\SQL\QueryObject;
 
 class Update implements QueryInterface
 {
